A few final renames and changes.
Stitch -> imageprocessor
automatebeta -> automate
