+++ imageprocessor.py
